Optimized table rerender on fetch
